Switch Generic Types to Unity's "Display Name"
I *think* this the change I'm looking for (I had compiler problems getting the source version of DeInspektor running, so wasn't able to test).  But for context:

When Unity is showing the names of array items, it will show the contents of the first serializable string field it finds in the object.  I regularly abuse this with [HideInInspector] "name" fields that are set from code, like:

https://www.dropbox.com/s/rd299467a2hl5cj/Screenshot%202017-01-23%2012.14.02.png?dl=0

https://www.dropbox.com/s/84hrmey3mf57bq7/Screenshot%202017-01-23%2012.14.08.png?dl=0

Which shows up as:

https://www.dropbox.com/s/h9avtl71nlbzmd2/Screenshot%202017-01-23%2012.12.46.png?dl=0

However, with DeInspektor, it's showing the type, so just:

https://www.dropbox.com/s/ig2f9fa8gjm7vdk/Screenshot%202017-01-23%2012.13.09.png?dl=0

Pretty sure this is the only change I need to get the Unity behavior, which would be great!